a lot

+++ app/vm.py
@@ -27,24 +27,25 @@ class VMC():
         sunstone = Sunstone(order.cluster)
 
         # if ceph ds, create datablock first
-        posts = '{"image":{"NAME":"%s","TYPE":"DATABLOCK","PERSISTENT":"YES","SIZE":"%s","FSTYPE":"raw"},"ds_id":"%s"}'
         if ds_mad == "ceph":
+            ready_tag = False
+            posts = '{"image":{"NAME":"%s","TYPE":"DATABLOCK","PERSISTENT":"YES","SIZE":"%s","FSTYPE":"raw"},"ds_id":"%s"}'
             for x in order.disk:
                 params = posts % (time.time(), x*1024, order.cluster.ds_id)
                 resp = sunstone.image(method='POST', params=params)
                 resp = json.loads(resp.data)
                 disk_arr.append(resp['IMAGE']['ID'])
 
-        ready_tag = False
-        while not ready_tag:
-            ready_tag = True
-            for d in disk_arr:
-                resp = sunstone.image_id(d)
-                data = json.loads(resp.data)
-                if data['IMAGE']['STATE'] != "1":
-                    ready_tag = False
+            while not ready_tag:
+                ready_tag = True
+                for x in disk_arr:
+                    resp = sunstone.image_id(x)
+                    data = json.loads(resp.data)
+                    if data['IMAGE']['STATE'] != "1":
+                        ready_tag = False
 
-        template = render_template('vm/occi_template.html', order=order, network_dict=network_dict, disk_arr=disk_arr)
+        # create vm
+        template = render_template('vm/occi_template.html', order=order, network_dict=network_dict, disk_arr=disk_arr, ds_mad=ds_mad)
         resp = occi.compute_create(template)
         if resp.status != 201:
             return 0
@@ -55,11 +56,13 @@ class VMC():
         state = xmldoc.getElementsByTagName('STATE')[0].firstChild.nodeValue
         ip = [a.firstChild.nodeValue for a in xmldoc.getElementsByTagName('IP')]
 
-        posts = '{"action":{"perform":"rename","params":{"name":"%s"}}}'
+        # rename ceph disk
         if ds_mad == "ceph":
+            posts = '{"action":{"perform":"rename","params":{"name":"%s"}}}'
             for index,d in enumerate(disk_arr):
                 resp = sunstone.image_action(d, (posts % ("%s-%s" % (vm_id, index+1))))
 
+        # write vm info to db
         try:
             vm = VM(hostname, ip, order.cpu, order.mem, order.network, order.disk, order.desc, order.id, 
                     order.cluster_id, vm_id, state, -1, order.user.id, 0, order.cluster.if_test, order.days)
@@ -70,6 +73,27 @@ class VMC():
             return 1
         except:
             return 0
+
+
+    def detail(self, vm):
+        ret = {}
+        sunstone = Sunstone(vm.cluster)
+        resp = sunstone.vm_id(vm.vm_id)
+        info = json.loads(resp.data)['VM']
+
+        nic      = info['TEMPLATE'].has_key('NIC') and info['TEMPLATE']['NIC'] or []
+        disk     = info['TEMPLATE'].has_key('DISK') and info['TEMPLATE']['DISK'] or []
+        snapshot = info['TEMPLATE'].has_key('SNAPSHOT') and info['TEMPLATE']['SNAPSHOT'] or []
+
+        nic      = type(nic)==dict and [nic] or nic
+        disk     = type(disk)==dict and [disk] or disk
+        snapshot = type(snapshot)==dict and [snapshot] or snapshot
+
+        ret['NIC'] = nic
+        ret['DISK'] = disk
+        ret['SNAPSHOT'] = snapshot
+
+        return ret
 
 
     def gen_price(self, vm):
@@ -87,10 +111,43 @@ class VMC():
 
 
     def rename(self, vm, newname):
-        sunstone = Sunstone(vm.order.cluster)
+        sunstone = Sunstone(vm.cluster)
         params = '{"action":{"perform":"rename","params":{"name":"%s"}}}' % newname
         resp = sunstone.vm_action(vm.vm_id, params)
-        return 1 if resp.status==204 else 0
+        return resp.status==204 and 1 or 0
+
+
+    def resize(self, vm, cpu, mem):
+        (cpu, mem) = (int(cpu), int(mem))
+        sunstone = Sunstone(vm.cluster)
+        params = '{"action":{"perform":"resize","params":{"vm_template":{"CPU":"%s","MEMORY":"%s"},"enforce":false}}}' % (cpu, mem*1024)
+        resp = sunstone.vm_action(vm.vm_id, params)
+        if resp.status == 204:
+            vm.cpu = cpu
+            vm.mem = mem
+            db.session.add(vm)
+            db.session.commit()
+            vminst = VMC()
+            vminst.gen_price(vm)
+            return 1
+        else:
+            return 0
+
+
+    def attach_nic(self, vm, network_id):
+        sunstone = Sunstone(vm.cluster)
+        network = Network.query.get(network_id)
+        params = '{"action":{"perform":"attachnic","params":{"nic_template":{"NIC":{"NETWORK":"%s","NETWORK_UNAME":"oneadmin","MODEL":"virtio"}}}}}' % \
+                  network.sunstone_name
+        resp = sunstone.vm_action(vm.vm_id, params)
+        return resp.status==204 and 1 or 0
+
+
+    def detach_nic(self, vm, nic_id):
+        sunstone = Sunstone(vm.cluster)
+        params = '{"action":{"perform":"detachnic","params":{"nic_id":"%s"}}}' % nic_id
+        resp = sunstone.vm_action(vm.vm_id, params)
+        return resp.status==204 and 1 or 0
 
 
     def refresh(self):
@@ -100,38 +157,89 @@ class VMC():
             resp = sunstone.vm()
             if resp.status != 200:
                 return 0
-            info = json.loads(resp.data)['VM_POOL']['VM']
-            if type(info) == dict:
-                info = [info]
+            try:
+                info = json.loads(resp.data)['VM_POOL']['VM']
+            except:
+                return 1
+            info = type(info)==dict and [info] or info
             for y in info:
                 vm = VM.query.filter(and_(VM.cluster_id==x.id, VM.vm_id==int(y['ID']))).first()
                 if not vm:
-                    next
+                    continue
+                #IP
+                nics = y['TEMPLATE'].has_key('NIC') and y['TEMPLATE']['NIC'] or []
+                nics = type(nics)==dict and [nics] or nics
+                ip = [z['IP'] for z in nics]
+                vm.ip = ' '.join(ip)
+                #Status
                 lcm_state_code = int(y['LCM_STATE'])
                 vm_state_code = int(y['STATE'])
                 if lcm_state_code == 0:
                     vm.status = app.config['VM_STATE'][vm_state_code]
                 else:
                     vm.status = app.config['LCM_STATE'][lcm_state_code]
-                db.session.add(x)
+
+                db.session.add(vm)
                 db.session.commit()
         return 1
 
 
+    def start(self, vm):
+        sunstone = Sunstone(vm.cluster)
+        post = '{"action":{"perform":"resume"}}'
+        resp = sunstone.vm_action(vm.vm_id, post)
+        return resp.status==204 and 1 or 0
+
+
+    def poweroff(self, vm):
+        sunstone = Sunstone(vm.cluster)
+        post = '{"action":{"perform":"poweroff","params":{"hard":false}}}'
+        resp = sunstone.vm_action(vm.vm_id, post)
+        return resp.status==204 and 1 or 0
+
+
+    def poweroff_hard(self, vm):
+        sunstone = Sunstone(vm.cluster)
+        post = '{"action":{"perform":"poweroff","params":{"hard":true}}}'
+        resp = sunstone.vm_action(vm.vm_id, post)
+        return resp.status==204 and 1 or 0
+
+
+    def reboot(self, vm):
+        sunstone = Sunstone(vm.cluster)
+        post = '{"action":{"perform":"reboot"}}'
+        resp = sunstone.vm_action(vm.vm_id, post)
+        return resp.status==204 and 1 or 0
+
+
+    def reboot_hard(self, vm):
+        sunstone = Sunstone(vm.cluster)
+        post = '{"action":{"perform":"reset"}}'
+        resp = sunstone.vm_action(vm.vm_id, post)
+        return resp.status==204 and 1 or 0
+
+
+    def recreate(self, vm):
+        sunstone = Sunstone(vm.cluster)
+        post = '{"action":{"perform":"resubmit"}}'
+        resp = sunstone.vm_action(vm.vm_id, post)
+        return resp.status==204 and 1 or 0
+
+
     def delete(self, vm):
-        occi = OCCI(vm.order.cluster)
-        sunstone = Sunstone(vm.order.cluster)
+        occi = OCCI(vm.cluster)
+        sunstone = Sunstone(vm.cluster)
         vm_info = json.loads(sunstone.vm_id(vm.vm_id).data)
 
         resp = occi.compute_delete(vm.vm_id)
         if resp.status != 204:
             return 0
 
-        disk = vm_info['VM']['TEMPLATE']['DISK']
+        disk = 'DISK' in vm_info['VM']['TEMPLATE'] and vm_info['VM']['TEMPLATE']['DISK'] or []
         if type(disk) == dict:
             disk = [disk]
         for x in disk:
-            if x['CLONE'] == 'NO':
+            if "CLONE" in x and x['CLONE'] == 'NO':
                 sunstone.image_id(image_id=x['IMAGE_ID'], method='DELETE')
 
         try:
@@ -140,4 +248,3 @@ class VMC():
             return 1
         except:
             return 0
-        return 1
